--- feature/links/components/index.tsx
@@ -1,10 +1,0 @@
-// Cards
-export { HorizontalCard } from "./cards/HorizontalCard";
-export { FeaturedLinksCard } from "./cards/FeaturedCard";
-
-// Lists
-export { FeaturedLinksList } from "./lists/FeaturedList";
-export { LinksFlatList } from "./lists/LinksFlatList";
-
-// View
-export { LinksTopView } from "../presentation/views/LinksTopView";
