+++ feature/links/presentation/components/display/cards/ErrorCard.tsx
