+++ feature/links/presentation/components/display/cards/SwipeFinishCard.tsx
