+++ feature/links/presentation/components/display/images/SwipeCardImage.tsx
