+++ feature/links/presentation/components/display/images/index.tsx
@@ -1,0 +1,1 @@
+export * from "./SwipeCardImage";
