+++ feature/links/presentation/components/display/lists/index.tsx
@@ -1,0 +1,2 @@
+export * from "./FeaturedList";
+export * from "./LinksFlatList";
