+++ feature/links/presentation/components/display/pagination/index.tsx
@@ -1,0 +1,1 @@
+export * from "./PaginationDots";
