+++ feature/links/presentation/components/display/preview/LinkPreview.tsx
@@ -1,9 +1,8 @@
 import { View } from "react-native";
 
 import { ThemedText } from "@/components/text/ThemedText";
-import type { OGData } from "../../domain/models/types/links";
-import { HorizontalCard } from "../cards/HorizontalCard";
-import { LoadingCard } from "../cards/LoadingCard";
+import type { OGData } from "../../../../domain/models/types";
+import { HorizontalCard, LoadingCard } from "../cards";
 
 interface LinkPreviewProps {
   url: string;
