+++ feature/links/presentation/components/input/form/index.tsx
@@ -1,0 +1,1 @@
+export * from "./linkInputForm";
