+++ feature/links/presentation/components/overlay/index.tsx
@@ -1,0 +1,1 @@
+export * from "../overlay/SwipeDirectionOverlay";
