+++ feature/links/presentation/screens/SwipeScreen.tsx
@@ -7,14 +7,16 @@ import Animated, {
 } from "react-native-reanimated";
 
 import { useGetLinks, useOGDataBatch } from "@/feature/links/application/hooks";
-import { SwipeActions } from "../../components/actions/SwipeActions";
-import { LinkInfoCard } from "../../components/cards/LinkInfoCard";
-import { SwipeFinishCard } from "../../components/cards/SwipeFinishCard";
-import { SwipeCardImage } from "../../components/images/SwipeCardImage";
-import { SwipeDirectionOverlay } from "../../components/overlay/SwipeDirectionOverlay";
-import { PaginationDots } from "../../components/pagination/PaginationDots";
-import { OVERLAY_BACKGROUND_COLORS } from "../../domain/constants";
-import { type Card } from "../../domain/models/types";
+import { OVERLAY_BACKGROUND_COLORS } from "@/feature/links/domain/constants";
+import { type Card } from "@/feature/links/domain/models/types";
+import {
+  LinkInfoCard,
+  PaginationDots,
+  SwipeFinishCard,
+} from "@/feature/links/presentation/components/display";
+import { SwipeCardImage } from "@/feature/links/presentation/components/display/images";
+import { SwipeActions } from "@/feature/links/presentation/components/input";
+import { SwipeDirectionOverlay } from "@/feature/links/presentation/components/overlay";
 
 export default function SwipeScreen() {
   const [isFinished, setIsFinished] = useState(false);
